+++ packages/react/.storybook/config.js
@@ -10,9 +10,7 @@ withOptions({
 });
 
 // load global styles
-// require('!style-loader!css-loader!@carbon/charts/demo/styles.css');
-// Temporary override for local style testing
-require('!style-loader!css-loader!sass-loader!../../core/src/styles/styles-white.scss');
+require('!style-loader!css-loader!@carbon/charts/demo/styles.css');
 
 const req = require.context('../', true, /.stories.js$/);
 function loadStories() {
