+++ packages/react/package.json
@@ -73,7 +73,6 @@
 		"base64-inline-loader": "1.1.0",
 		"carbon-components-react": "7.36.0",
 		"css-loader": "0.28.7",
-		"elkjs": "0.7.1",
 		"img-loader": "2.0.0",
 		"react": "17.0.2",
 		"react-dom": "17.0.2",
