--- packages/react/src/components/diagrams/Diagrams.stories.js
@@ -1,153 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import CardNode, {
-	CardNodeColumn,
-	CardNodeTitle,
-	CardNodeSubtitle
-} from './CardNode';
-import Edge from './Edge';
-import ShapeNode from './ShapeNode';
-import {ArrowRightMarker} from './Marker';
-
-import { User16, Wikis16, Debug16 } from '@carbon/icons-react';
-import { elbow } from './Edge/buildPath';
-
-const nodeHeight = 64;
-const nodeWidth = 200;
-const ShapeNodeSize = 64;
-
-const stories = storiesOf('Diagrams|Graph', module);
-stories.addDecorator((story) => (
-	<div className="container theme--white">{story()}</div>
-));
-
-stories.add('Composed', () => {
-	return (
-		<svg height="1000" width="1000">
-			<defs>
-				<ArrowRightMarker id={"marker"} />
-			</defs>
-			<Edge
-				source={{ x: 200, y: 132 }}
-				target={{ x: 596, y: 132 }}
-				markerEnd={"marker"}
-				variant={'dash-sm'}
-			/>
-
-			<foreignObject
-				style={{ overflow: 'visible' }}
-				transform={`translate(${100},${100})`}
-				height={nodeHeight}
-				width={nodeWidth}>
-				<CardNode>
-					<CardNodeColumn>
-						<User16 />
-					</CardNodeColumn>
-					<CardNodeColumn>
-						<CardNodeTitle>Title</CardNodeTitle>
-						<CardNodeSubtitle>Description</CardNodeSubtitle>
-					</CardNodeColumn>
-				</CardNode>
-			</foreignObject>
-
-			<foreignObject
-				style={{ overflow: 'visible' }}
-				transform={`translate(${600},${100})`}>
-				<ShapeNode title={'Title'} size={ShapeNodeSize} renderIcon={<Wikis16 />} />
-			</foreignObject>
-		</svg>
-	);
-});
-
-stories.add('Programmatic', () => {
-	const nodeData = [
-		{ id: 'a', x: 0, y: 0, icon: <User16 />, nodeWidth, nodeHeight },
-		{ id: 'b', x: 250, y: 0, icon: <Wikis16 />, nodeWidth, nodeHeight },
-		{
-			id: 'c',
-			x: 600,
-			y: 200,
-			icon: <Debug16 />,
-			ShapeNode: true,
-			nodeWidth: ShapeNodeSize,
-			nodeHeight: ShapeNodeSize,
-		},
-		{ id: 'd', x: 20, y: 280, icon: <Wikis16 />, nodeWidth, nodeHeight },
-	];
-
-	const edgeData = [
-		{ source: 'a', target: 'b', variant: 'dash-md' },
-		{
-			source: 'c',
-			target: 'b',
-			path: (source, target) => elbow(source, target),
-		},
-		{
-			source: 'd',
-			target: 'c',
-			path: (source, target) => elbow(source, target),
-			variant: 'tunnel',
-		},
-	];
-
-	const edgeMapped = edgeData.map((link) => {
-		const sourceNode = nodeData.find((node) => node.id === link.source);
-		const targetNode = nodeData.find((node) => node.id === link.target);
-
-		return {
-			...link,
-			source: {
-				x: sourceNode.x + sourceNode.nodeWidth / 2,
-				y: sourceNode.y + sourceNode.nodeHeight / 2,
-			},
-			target: {
-				x: targetNode.x + targetNode.nodeWidth / 2,
-				y: targetNode.y + targetNode.nodeHeight / 2,
-			},
-		};
-	});
-
-	const nodes = nodeData.map((node, i) => (
-		<foreignObject
-			style={{ overflow: 'visible' }}
-			key={`node_${i}`}
-			transform={`translate(${node.x},${node.y})`}
-			height={node.nodeHeight}
-			width={node.nodeWidth}>
-			{node.ShapeNode ? (
-				<ShapeNode
-					title={'Title'}
-					size={ShapeNodeSize}
-					description={'Description'}
-					renderIcon={node.icon}
-				/>
-			) : (
-				<CardNode>
-					<CardNodeColumn>
-						{node.icon}
-					</CardNodeColumn>
-					<CardNodeColumn>
-						<CardNodeTitle>Title</CardNodeTitle>
-						<CardNodeSubtitle>Description</CardNodeSubtitle>
-					</CardNodeColumn>
-				</CardNode>
-			)}
-		</foreignObject>
-	));
-	const edges = edgeMapped.map((edge, i) => (
-		<Edge
-			key={`link_${i}`}
-			source={edge.source}
-			target={edge.target}
-			path={edge.path && edge.path(edge.source, edge.target)}
-			variant={edge.variant}
-		/>
-	));
-
-	return (
-		<svg height="1000" width="1000">
-			{edges}
-			{nodes}
-		</svg>
-	);
-});
